requirements.txt
jinja2==2.9.6
pyyaml==3.12
six==1.10.0
